+++ misc.py
@@ -37,7 +37,7 @@ def col3to1(threecol):
 
     # dummy, N = threecol.shape
 
-    # onecol = numpy.zeros(3*N, dtype=numpy.complex64)
+    # onecol = numpy.zeros(3*N, dtype=numpy.complex128)
     # ind = numpy.multiply(range(2, N), 3)
     # onecol[ind-2] = threecol[:, ]
     # onecol[ind-1] = threecol[:, 2]
@@ -187,7 +187,7 @@ def read_data(filename):
         n_rows = len(lines)
         n_cols = len(lines[0].split(','))
 
-        dat = numpy.zeros([n_rows, n_cols], dtype=numpy.complex64)
+        dat = numpy.zeros([n_rows, n_cols], dtype=numpy.complex128)
         for i in range(n_rows):
             l = lines[i].split(',')
             for j in range(n_cols):
